V3.83 (#475)
* Updating SEMVER to v3.83

* code hygiene

* WIP: adding Bessel functions to the library

* WIP: Bessel function evaluations

* WIP: exploring a mixed-precision Conv2D operator architecture

* WIP: second mixed-precision Conv2D operator architecture

* Mark blfloat non-tmpl functions as inline (#474)

Fixes odr linking issues with multiple TU

* header edits

* restructuring bfloat to new directory structure for mathlib and adding sqrt test

* code hygiene on bfloat16

* WIP: adding min/max shim to bfloat16 mathlib

* WIP: adding bfloat16 shim for hypot

* WIP: adding hyperbolic functions to bfloat mathlib shim

* Interesting: bfloat16 and float16 are too small to faithfully represent the hypot intermediate so we need to go through double

* enhancing the generic test suite to automatically clip when there are too many errors

* adding trig functions

* adding logarithmic functions

* adding exponent functions to bfloat mathlib shim

* bug fix: GCC doesn't support hypotl in its math lib, so marshalling through double

* Adding truncation functions and bug fixes on isnan and isinf

* WIP: trying to get GCC back into the fold: adding math.h C math lib header

* Adding fmod/remainder to the bfloat mathlib

* RCA'ed the hyperbolic cotangent and hyperbolic secant ULP errors: need to compute through float

* Adding fpclassify implementation for bfloat mathlib

* Adding error and gamma functions to the bfloat mathlib shim

* WIP: trying to isolate the CI failure that doesn't reproduce locally on G++ 13.3.0

* WIP: updating the actions/upload-artifact@v4

* WIP: prepping for Linux, Vcpkg, and Conda package release engineering

* restructuring cfloat mathlib and generalizing the mathlib test suite

* Adding integer power to bfloat mathlib shim

* adding high precision constexpr constants to cfloat and long double

* adding bfloat16 perf benchmark

* code hygiene and disabling constexpr for long double

* Adding hw float performance assessment to bfloat16 benchmark

* PR review fix of bad SFINAE pattern

* fixing documentation error for atan/acos/asin where were mislabeled

* consistency fix: using the same pattern for accumulating test case failures for fpclassify

* renaming step 1

* renaming step 2

* rename step 3

* rename step 4

* updating CMakeLists.txt to point to correct bfloat16 directory

* fixing dd constants for sqrt(3) and sqrt(5)

* bug fixes

* dd constant reference fix

* bug fix of bfloat16::isinteger(): need to filter out nan and inf

---------

+++ include/sw/universal/verification/test_suite.hpp
@@ -117,7 +117,7 @@ namespace sw {
 		void ExplicitConversions(TestType a) {
 			std::cout << "Explicit conversions for " << type_tag(a) << '\n';
 			std::cout << to_binary(a) << " : " << a << '\n';
-			uint8_t u8 = char(a); // conversion to uint8_t
+			uint8_t u8 = uint8_t(char(a)); // conversion to uint8_t
 			std::cout << "uint8_t  conversion : " << to_binary(u8) << " : " << u8 << '\n';
 			uint16_t u16 = (unsigned short)(a); // conversion to uint16_t
 			std::cout << "uint16_t conversion : " << to_binary(u16) << " : " << u16 << '\n';
